Rename sync files
